Change SignMixin to release GIL when calling C_Sign/C_SignFinal
Release the Python global interpreter lock while making calls to the
PKCS#11 C_Sign and C_SignFinal functions, to allow other Python threads
to run while a sign operation is in progress. This is helpful when the
sign call blocks waiting for user interaction, such as the user needing
to touch a security key.