_pkcs11.pyx: Fix Python int too large to convert to C ssize_t
When a key is retrieved it can retrun CK_UNAVAILABLE_INFORMATION for
some of its properties. The CK_UNAVAILABLE_INFORMATION value is set to
~0UL or the largest int possible on a system. The python-pkcs11 library
would then try to create a buffer of that size and crash.

This commit adds error checking to make sure when
CK_UNAVAILABLE_INFORMATION is returned that the property value is set to
blank and returned immediately.